Put Transformations for SpinerEOSDependsRhoSie in seperate header (eos_spiner_sie_transfroms.hpp)
